api/requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
sqlalchemy==2.0.43
psycopg[binary]==3.2.9
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.10.1
python-multipart==0.0.20
email-validator==2.2.0
httpx==0.28.1
